+++ output_item.c
@@ -6,7 +6,7 @@
 /*   By: terabu <terabu@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/01/30 11:37:37 by terabu            #+#    #+#             */
-/*   Updated: 2023/01/31 09:45:29 by terabu           ###   ########.fr       */
+/*   Updated: 2023/02/01 11:24:00 by terabu           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 void	output_item(t_solong *solong)
 {
-	int	i;
-	int	j;
+	int		i;
+	size_t	j;
 
 	i = 0;
 	while (i < solong->row)
